# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hangout Alley Siargao Restaurant Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MENU ITEMS
  /menu:
    get:
      operationId: getMenuItems
      tags: [menu]
      summary: List all menu items
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: available
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      summary: Create a new menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      summary: Get a menu item by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "404":
          description: Not found
    put:
      operationId: updateMenuItem
      tags: [menu]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuItemBody"
      responses:
        "200":
          description: Updated menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /inventory:
    get:
      operationId: getInventory
      tags: [inventory]
      summary: List all inventory items
      responses:
        "200":
          description: List of inventory items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryItem"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/{id}:
    put:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInventoryItemBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
      required: [id, name, sortOrder, createdAt]
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required: [name, sortOrder]
    MenuItem:
      type: object
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        allergens:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, name, price, isAvailable, isFeatured, createdAt]
    CreateMenuItemBody:
      type: object
      properties:
        categoryId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isFeatured:
          type: boolean
        allergens:
          type: ["string", "null"]
      required: [name, price, isAvailable, isFeatured]
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        reorderLevel:
          type: number
        costPerUnit:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
        updatedAt:
          type: string
      required: [id, name, unit, currentStock, reorderLevel, updatedAt]
    CreateInventoryItemBody:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        currentStock:
          type: number
        reorderLevel:
          type: number
        costPerUnit:
          type: ["number", "null"]
        supplier:
          type: ["string", "null"]
      required: [name, unit, currentStock, reorderLevel]
...